The changes look comprehensive and well-structured. You've successfully:
1. Created a centralized `prediction_utils.py` with common model-related functions
2. Removed redundant code from `evaluation.py`, `prediction.py`, and `training_utils.py`
3. Updated import statements and function calls across files
4. Maintained the existing functionality while improving code organization

The refactoring achieves several key goals:
- Reduces code duplication
- Centralizes common model-related utilities
- Improves maintainability
- Keeps the code DRY (Don't Repeat Yourself)

A few observations and potential improvements:
- The new `prediction_utils.py` looks clean and well-documented
- The function signatures maintain compatibility with existing code
- Error handling and logging are preserved

Would you like me to review any specific aspect of the changes or suggest any further refinements?
