+++ autoslide/pipeline/model/prediction.py
@@ -10,25 +10,15 @@ This script:
 
 from autoslide.pipeline import utils
 import os
-import slideio
-from matplotlib import pyplot as plt
-import cv2 as cv
 import numpy as np
-from pprint import pprint
-import pandas as pd
-from skimage import morphology as morph
-from scipy.ndimage import binary_fill_holes
 from glob import glob
-import torch
-import torchvision
-from torchvision import transforms as T
-from torchvision.models.detection.faster_rcnn import FastRCNNPredictor
-from torchvision.models.detection.mask_rcnn import MaskRCNNPredictor
 from PIL import Image
-from importlib import reload
-from autoslide import config
+import matplotlib.pyplot as plt
 from tqdm import tqdm
 import argparse
+
+from autoslide import config
+from autoslide.pipeline.model.prediction_utils import load_model, predict_single_image
 
 # Get directories from config
 data_dir = config['data_dir']
@@ -36,43 +26,6 @@ artifacts_dir = config['artifacts_dir']
 plot_dir = config['plot_dirs']
 
 
-def load_model(model_path=None):
-    """
-    Load the trained Mask R-CNN model.
-    
-    Args:
-        model_path (str): Path to the saved model. If None, uses default path.
-        
-    Returns:
-        tuple: (model, device, transform) - Loaded model, device, and transform
-    """
-    if model_path is None:
-        model_path = os.path.join(artifacts_dir, 'best_val_mask_rcnn_model.pth')
-    
-    if not os.path.exists(model_path):
-        raise FileNotFoundError(f"Model file not found at {model_path}")
-    
-    # Recreate base model
-    model = torchvision.models.detection.maskrcnn_resnet50_fpn()
-    in_features = model.roi_heads.box_predictor.cls_score.in_features
-    model.roi_heads.box_predictor = FastRCNNPredictor(in_features, 2)
-    in_features_mask = model.roi_heads.mask_predictor.conv5_mask.in_channels
-    hidden_layer = 256
-    model.roi_heads.mask_predictor = MaskRCNNPredictor(
-        in_features_mask, hidden_layer, 2)
-    
-    # Load model weights
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    model.load_state_dict(torch.load(model_path, map_location=device))
-    model.eval()
-    model.to(device)
-    
-    transform = T.ToTensor()
-    
-    print(f"Model loaded from: {model_path}")
-    print(f"Using device: {device}")
-    
-    return model, device, transform
 
 
 def find_images_to_process():
@@ -120,9 +73,9 @@ def find_images_to_process():
     return images_to_process
 
 
-def predict_single_image(model, image_path, device, transform):
+def predict_image_from_path(model, image_path, device, transform):
     """
-    Perform prediction on a single image.
+    Perform prediction on a single image from file path.
     
     Args:
         model: Trained Mask R-CNN model
@@ -134,46 +87,7 @@ def predict_single_image(model, image_path, device, transform):
         numpy.ndarray: Combined predicted mask
     """
     try:
-        # Load and preprocess image
-        img = Image.open(image_path).convert("RGB")
-        img_tensor = transform(img).to(device)
-        
-        # Perform prediction
-        with torch.no_grad():
-            pred = model([img_tensor])
-        
-        # Combine all predicted masks
-        n_preds = len(pred[0]["masks"])
-        if n_preds > 0:
-            # Stack all predictions and take the mean
-            all_preds = np.stack([
-                (pred[0]["masks"][i].cpu().detach().numpy() * 255).astype("uint8").squeeze()
-                for i in range(n_preds)
-            ])
-            
-            # Weight by confidence scores if available
-            if len(pred[0]["scores"]) > 0:
-                scores = pred[0]["scores"].cpu().detach().numpy()
-                weighted_mask = np.zeros_like(all_preds[0], dtype=np.float32)
-                total_weight = 0
-                
-                for mask, score in zip(all_preds, scores):
-                    weighted_mask += mask.astype(np.float32) * score
-                    total_weight += score
-                
-                if total_weight > 0:
-                    combined_mask = (weighted_mask / total_weight).astype(np.uint8)
-                else:
-                    combined_mask = all_preds.mean(axis=0).astype(np.uint8)
-            else:
-                combined_mask = all_preds.mean(axis=0).astype(np.uint8)
-        else:
-            # No predictions found
-            img_array = np.array(img)
-            combined_mask = np.zeros((img_array.shape[0], img_array.shape[1]), dtype=np.uint8)
-        
-        return combined_mask
-        
+        return predict_single_image(model, image_path, device, transform, return_time=False)
     except Exception as e:
         print(f"Error processing {image_path}: {e}")
         return None
@@ -243,6 +157,8 @@ def process_all_images(model_path=None, save_visualizations=False, max_images=No
     
     # Load model
     model, device, transform = load_model(model_path)
+    print(f"Model loaded successfully")
+    print(f"Using device: {device}")
     
     # Find images to process
     images_to_process = find_images_to_process()
@@ -266,7 +182,7 @@ def process_all_images(model_path=None, save_visualizations=False, max_images=No
         image_name = item['image_name']
         
         # Perform prediction
-        predicted_mask = predict_single_image(model, image_path, device, transform)
+        predicted_mask = predict_image_from_path(model, image_path, device, transform)
         
         if predicted_mask is not None:
             try:
